backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
supabase
python-dotenv
line-bot-sdk
requests
aiohttp
pydantic[email]